Refactor: extract session lifecycle from client.go into auth.go
