+++ client.go
@@ -3,7 +3,6 @@ package filemaker
 import (
 	"bytes"
 	"context"
-	"encoding/base64"
 	"encoding/json"
 	"errors"
 	"fmt"
@@ -156,70 +155,6 @@ func New(host, database, username, password string, opts ...Option) (*Client, er
 		location:             cfg.location,
 		authSem:              make(chan struct{}, 1),
 	}, nil
-}
-
-// Authenticate eagerly establishes a session by logging in to the host, so a
-// caller can surface credential or connectivity errors at a chosen point rather
-// than on the first operation. It is optional: every operation authenticates
-// lazily on first use when no session exists yet.
-//
-// Calling it always performs a fresh login and replaces the stored token. It
-// does not log out an existing session first: the previous token is abandoned,
-// not invalidated, and lingers on the host until it times out. Call Logout
-// before Authenticate if you need the old session torn down promptly.
-// Concurrent calls collapse to a single login, so a burst of callers does not
-// produce a burst of sessions.
-func (c *Client) Authenticate(ctx context.Context) error {
-	c.mu.RLock()
-	used := c.token
-	c.mu.RUnlock()
-	return c.authenticate(ctx, used)
-}
-
-// ensureAuthenticated acquires the initial session token on first use (lazy
-// authentication). It is a cheap read-and-return once a session exists;
-// concurrent first-use callers collapse to a single login via authenticate's
-// observed-token de-duplication (an empty token is the de-dup key).
-func (c *Client) ensureAuthenticated(ctx context.Context) error {
-	c.mu.RLock()
-	token := c.token
-	c.mu.RUnlock()
-	if token != "" {
-		return nil
-	}
-	return c.authenticate(ctx, "")
-}
-
-// Logout ends the current database session, invalidating the token on the
-// host. The client remains usable afterward: a subsequent operation (or
-// Authenticate) establishes a fresh session. Logout is a no-op when no session
-// has been established yet.
-func (c *Client) Logout(ctx context.Context) error {
-	c.mu.RLock()
-	token := c.token
-	c.mu.RUnlock()
-	if token == "" {
-		return nil
-	}
-
-	// Logout identifies the session by the token in the URL and takes no
-	// Authorization header, so it bypasses the authed do() path.
-	req, err := http.NewRequestWithContext(ctx, http.MethodDelete, c.baseURL()+"/sessions/"+token, nil)
-	if err != nil {
-		return fmt.Errorf("filemaker: failed to build request: %w", err)
-	}
-	req.Header.Set("Content-Type", "application/json")
-
-	var rb responseBody
-	if err := c.send(req, &rb); err != nil {
-		return err
-	}
-
-	c.mu.Lock()
-	c.token = ""
-	c.lastActivity = time.Now()
-	c.mu.Unlock()
-	return nil
 }
 
 // LastActivity returns the time of the last successful request made with the
@@ -394,72 +329,6 @@ func (c *Client) send(req *http.Request, out *responseBody) error {
 	return out.check()
 }
 
-// login performs a Basic-auth login and returns a fresh session token. The
-// session is database-scoped, so both database and username are required here;
-// New defers their validation to this point so a credential-free client can
-// still reach the unauthenticated ProductInfo endpoint.
-func (c *Client) login(ctx context.Context) (string, error) {
-	switch {
-	case c.database == "":
-		return "", errors.New("filemaker: no database specified")
-	case c.username == "":
-		return "", errors.New("filemaker: no username specified")
-	}
-
-	req, err := http.NewRequestWithContext(ctx, http.MethodPost, c.baseURL()+"/sessions", bytes.NewReader([]byte("{}")))
-	if err != nil {
-		return "", fmt.Errorf("filemaker: failed to build request: %w", err)
-	}
-	req.Header.Set("Content-Type", "application/json")
-	req.Header.Set("Authorization", "Basic "+basicAuth(c.username, c.password))
-
-	var rb responseBody
-	if err := c.send(req, &rb); err != nil {
-		return "", err
-	}
-	if rb.Response.Token == "" {
-		return "", errors.New("filemaker: authentication response contained no token")
-	}
-	return rb.Response.Token, nil
-}
-
-// authenticate replaces the stored token with one from a fresh login,
-// de-duplicating concurrent callers. observedToken is the token the caller last
-// saw; if another goroutine already refreshed the token since then, this returns
-// without a network round-trip. It serves both first-time and repeat
-// authentication — the empty string as observedToken is the first-use key.
-//
-// authSem (a cap-1 channel used as a context-aware mutex) serializes logins so
-// concurrent callers collapse to one network call; a caller waiting on it honors
-// its ctx and bails on cancellation rather than blocking on an in-flight login.
-// The login itself runs without holding the read/write lock, so token reads (and
-// LastActivity) stay live during a refresh.
-func (c *Client) authenticate(ctx context.Context, observedToken string) error {
-	select {
-	case c.authSem <- struct{}{}:
-	case <-ctx.Done():
-		return ctx.Err()
-	}
-	defer func() { <-c.authSem }()
-
-	c.mu.RLock()
-	current := c.token
-	c.mu.RUnlock()
-	if current != observedToken {
-		return nil
-	}
-
-	token, err := c.login(ctx)
-	if err != nil {
-		return err
-	}
-	c.mu.Lock()
-	c.token = token
-	c.lastActivity = time.Now()
-	c.mu.Unlock()
-	return nil
-}
-
 // apiURL builds the version-scoped root of the Data API URL, above the database
 // scope. Host-level metadata endpoints such as productInfo hang off it directly
 // rather than under a database.
@@ -496,9 +365,4 @@ func normalizeHost(host string, allowInsecureHTTP bool) (string, error) {
 	}
 
 	return host, nil
-}
-
-// basicAuth builds the value for an HTTP Basic Authorization header.
-func basicAuth(username, password string) string {
-	return base64.StdEncoding.EncodeToString([]byte(username + ":" + password))
 }
